Snowy bot classes: interact with user and implement commands typed by user
Snowy is initialized without greeting the user and no assertions are used for checking possible null cases for input.

Creating a dialog box object with greeting message when snowy object is initialized. Adding assertions for empty input, empty task list, empty task.

After a snowy object is initialized, the first dialog box appears is the greeting box. Assertions help check for flaws in the code, as empty inputshould not be parsed and empty task should not be added to a task list and an empty task list should not be saved locally.